added dev-java/libreadline-java-0.8.0-r4
This version supports being build with dev-libs/libedit.

There is an open bug at Gentoo's Bugzilla for this at #208285
Here is the original bug-text:

----
I noticed, that the current ebuild for dev-java/libreadline-java does
only build the readline-part, but editline or plain-getline-modes are
available too.
They are handy when one wants to circumvent the GPL of readline or don't
want to use a special library at all.
----

Please note that plain-getline does not work anymore and is removed for
the time being.