added net-mail/freepops
Freepops is scheduled for removal from Gentoo's official tree for valid
reasons (upstream is dead, failes to compile, maintainer retired). See
Gentoo's Bugzilla #370459

At leas the "failes to compile" can be fixed rather easy, as it seems
only a change in a lua-header's comments is the cause. I fixed just that
error. Please note that the way those headers are parsed, is still quite
error-prone, so it might fail again with the next version of lua!

This would fix Gentoo's Bug #354865 and probably #205442 too.

As upstream is dead, is currently searching for a new maintainer and
provides no public bug-tracking, this patch was not sent upstream.